+++ src/components/MenuItem.js
@@ -4,7 +4,8 @@ import {
     View, Pressable,
 } from 'react-native';
 import Text from './Text';
-import styles, {getButtonBackgroundColorStyle, getIconFillColor} from '../styles/styles';
+import styles from '../styles/styles';
+import * as StyleUtils from '../styles/StyleUtils';
 import Icon from './Icon';
 import * as Expensicons from './Icon/Expensicons';
 import getButtonState from '../libs/getButtonState';
@@ -48,7 +49,7 @@ const MenuItem = props => (
         }}
         style={({hovered, pressed}) => ([
             styles.popoverMenuItem,
-            getButtonBackgroundColorStyle(getButtonState(props.focused || hovered, pressed, props.success, props.disabled, props.interactive)),
+            StyleUtils.getButtonBackgroundColorStyle(getButtonState(props.focused || hovered, pressed, props.success, props.disabled, props.interactive)),
             ..._.isArray(props.wrapperStyle) ? props.wrapperStyle : [props.wrapperStyle],
         ])}
         disabled={props.disabled}
@@ -67,7 +68,7 @@ const MenuItem = props => (
                                 src={props.icon}
                                 width={props.iconWidth}
                                 height={props.iconHeight}
-                                fill={props.iconFill || getIconFillColor(
+                                fill={props.iconFill || StyleUtils.getIconFillColor(
                                     getButtonState(props.focused || hovered, pressed, props.success, props.disabled, props.interactive),
                                 )}
                             />
@@ -119,7 +120,7 @@ const MenuItem = props => (
                         <View style={styles.popoverMenuIcon}>
                             <Icon
                                 src={props.iconRight}
-                                fill={getIconFillColor(getButtonState(props.focused || hovered, pressed, props.success, props.disabled, props.interactive))}
+                                fill={StyleUtils.getIconFillColor(getButtonState(props.focused || hovered, pressed, props.success, props.disabled, props.interactive))}
                             />
                         </View>
                     )}
