+++ src/components/ThumbnailImage.js
@@ -4,7 +4,8 @@ import {View} from 'react-native';
 import PropTypes from 'prop-types';
 import ImageWithSizeCalculation from './ImageWithSizeCalculation';
 import addEncryptedAuthTokenToURL from '../libs/addEncryptedAuthTokenToURL';
-import styles, {getWidthAndHeightStyle} from '../styles/styles';
+import styles from '../styles/styles';
+import * as StyleUtils from '../styles/StyleUtils';
 
 const propTypes = {
     /** Source URL for the preview image */
@@ -52,7 +53,7 @@ class ThumbnailImage extends PureComponent {
         return (
             <View
                 style={[
-                    getWidthAndHeightStyle(this.state.thumbnailWidth, this.state.thumbnailHeight),
+                    StyleUtils.getWidthAndHeightStyle(this.state.thumbnailWidth, this.state.thumbnailHeight),
                     styles.alignItemsCenter,
                     styles.justifyContentCenter,
                     styles.overflowHidden,
