+++ src/pages/home/report/EmojiPickerMenuItem.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {Pressable} from 'react-native';
-import styles, {getButtonBackgroundColorStyle} from '../../../styles/styles';
+import styles from '../../../styles/styles';
+import * as StyleUtils from '../../../styles/StyleUtils';
 import getButtonState from '../../../libs/getButtonState';
 import Hoverable from '../../../components/Hoverable';
 import Text from '../../../components/Text';
@@ -27,7 +28,7 @@ const EmojiPickerMenuItem = props => (
             pressed,
         }) => ([
             styles.pv1,
-            getButtonBackgroundColorStyle(getButtonState(false, pressed)),
+            StyleUtils.getButtonBackgroundColorStyle(getButtonState(false, pressed)),
             props.isHighlighted ? styles.emojiItemHighlighted : {},
             styles.emojiItem,
         ])}
